Changed hostname template, added vm host
I changed the hostnames because I really wanted to make the system
more flexible than what it was before. This new naming scheme makes
it easier for me to find out where devices are and use proper names
for where these devices are (instead of trying to locate cities in
which AWS has datacenters or other companies, which is foolish).

I also added a new VM host to the VARS network last week.